+++ crypto/rand/default.go
@@ -7,7 +7,6 @@ package rand
 import (
 	"io"
 	"math/big"
-	"sync"
 	"time"
 )
 
@@ -18,27 +17,7 @@ func Reader() io.Reader {
 	return globalRand
 }
 
-type lockingPRNG struct {
-	*PRNG
-	mu sync.Mutex
-}
-
 var globalRand *lockingPRNG
-
-func init() {
-	p, err := NewPRNG()
-	if err != nil {
-		panic(err)
-	}
-	globalRand = &lockingPRNG{PRNG: p}
-}
-
-func (p *lockingPRNG) Read(s []byte) (n int, err error) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-
-	return p.PRNG.Read(s)
-}
 
 // Read fills b with random bytes obtained from the default userspace PRNG.
 func Read(b []byte) {
@@ -48,32 +27,32 @@ func Read(b []byte) {
 
 // Uint32 returns a uniform random uint32.
 func Uint32() uint32 {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Uint32()
 }
 
 // Uint64 returns a uniform random uint64.
 func Uint64() uint64 {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Uint64()
 }
 
 // Uint32N returns a random uint32 in range [0,n) without modulo bias.
 func Uint32N(n uint32) uint32 {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Uint32N(n)
 }
 
 // Uint64N returns a random uint32 in range [0,n) without modulo bias.
 func Uint64N(n uint64) uint64 {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Uint64N(n)
 }
@@ -81,8 +60,8 @@ func Uint64N(n uint64) uint64 {
 // Int32 returns a random 31-bit non-negative integer as an int32 without
 // modulo bias.
 func Int32() int32 {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Int32()
 }
@@ -91,8 +70,8 @@ func Int32() int32 {
 // without modulo bias.
 // Panics if n <= 0.
 func Int32N(n int32) int32 {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Int32N(n)
 }
@@ -100,8 +79,8 @@ func Int32N(n int32) int32 {
 // Int64 returns a random 63-bit non-negative integer as an int64 without
 // modulo bias.
 func Int64() int64 {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Int64()
 }
@@ -110,16 +89,16 @@ func Int64() int64 {
 // without modulo bias.
 // Panics if n <= 0.
 func Int64N(n int64) int64 {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Int64N(n)
 }
 
 // Int returns a non-negative integer without bias.
 func Int() int {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Int()
 }
@@ -128,16 +107,16 @@ func Int() int {
 // modulo bias.
 // Panics if n <= 0.
 func IntN(n int) int {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.IntN(n)
 }
 
 // UintN returns, as an uint, a random integer in [0,n) without modulo bias.
 func UintN(n uint) uint {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.UintN(n)
 }
@@ -145,8 +124,8 @@ func UintN(n uint) uint {
 // Duration returns a random duration in [0,n) without modulo bias.
 // Panics if n <= 0.
 func Duration(n time.Duration) time.Duration {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.Duration(n)
 }
@@ -155,8 +134,8 @@ func Duration(n time.Duration) time.Duration {
 // indexes i and j.
 // Panics if n < 0.
 func Shuffle(n int, swap func(i, j int)) {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	globalRand.Shuffle(n, swap)
 }
@@ -171,8 +150,8 @@ func ShuffleSlice[S ~[]E, E any](s S) {
 // Int returns a uniform random value in [0,max).
 // Panics if max <= 0.
 func BigInt(max *big.Int) *big.Int {
-	globalRand.mu.Lock()
-	defer globalRand.mu.Unlock()
+	globalRand.Lock()
+	defer globalRand.Unlock()
 
 	return globalRand.PRNG.BigInt(max)
 }
